# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "price",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "stock",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "moreOrequal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lessOrEqual",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductRespondeDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      }
    },
...
    "/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderRespondeDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ]
      },
      "patch": {
        "operationId": "OrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/resport/{id}": {
      "get": {
        "operationId": "OrdersController_generatePdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginUserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/recovery-password": {
      "post": {
        "operationId": "AuthController_generateRecovery",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryPasswordDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryPasswordDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/create-student": {
      "post": {
        "operationId": "AuthController_createStudent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/recovery-password/{recovery_token}": {
      "get": {
        "operationId": "AuthController_getRecoveryById",
        "parameters": [
          {
            "name": "recovery_token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryPasswordResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      },
      "post": {
        "operationId": "AuthController_postRecoveryById",
        "parameters": [
          {
            "name": "recovery_token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryPasswordResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/change-password": {
      "put": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    }
  },
  "info": {
    "title": "MI PROFE",
    "description": "The MI PROFE-API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "api"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "idDocument": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "birthdate": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "email",
          "role",
          "idDocument",
          "lastName",
          "firstName",
          "status",
          "birthdate"
        ]
      },
...
      "IdCreateEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          }
        }
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stock": {
            "type": "number"
          },
          "status": {
            "type": "boolean"
          },
          "category": {
            "$ref": "#/components/schemas/IdCreateEntity"
          }
        },
        "required": [
          "name",
          "price",
          "stock",
          "status",
          "category"
        ]
      },
      "Category": {
        "type": "object",
        "properties": {}
      },
      "ProductRespondeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "status": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stock": {
            "type": "number"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          }
        },
        "required": [
          "id",
          "status",
          "name",
          "price",
          "stock",
          "category"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stock": {
            "type": "number"
          },
          "status": {
            "type": "boolean"
          },
          "category": {
            "$ref": "#/components/schemas/IdCreateEntity"
          }
        }
      },
...
      "ReportsResponseDto": {
        "type": "object",
        "properties": {
          "reportUrl": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "reportUrl",
          "name"
        ]
      },
...
      "CreateOrderProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "amount": {
            "type": "number"
          },
          "price": {
            "type": "number"
          },
          "subtotal": {
            "type": "number"
          },
          "product": {
            "$ref": "#/components/schemas/IdCreateEntity"
          }
        },
        "required": [
          "amount",
          "price",
          "subtotal",
          "product"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "deadline": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "orderProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderProductDto"
            }
          }
        },
        "required": [
          "date",
          "stage",
          "provider",
          "total",
          "orderProducts"
        ]
      },
      "OrderRespondeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "deadline": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "provider": {
            "type": "string"
          },
          "orderProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderProductDto"
            }
          }
        },
        "required": [
          "id",
          "date",
          "deadline",
          "stage",
          "total",
          "provider",
          "orderProducts"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "deadline": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "orderProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderProductDto"
            }
          }
        }
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginUserResponseDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "number"
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "idDocument": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "loginStamp": {
            "type": "number"
          }
        },
        "required": [
          "email",
          "id",
          "lastName",
          "firstName",
          "idDocument",
          "role",
          "loginStamp"
        ]
      },
      "RecoveryPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "RecoveryPasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "newPassword"
        ]
      },
      "ChangePasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      }
    }
  }
}